refactor/관리 패키지 위치 변경
