+++ app/src/main/java/org/sopt/at/views/signin/LoginFieldType.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.models
+package org.sopt.at.views.signin
 
 enum class LoginFieldType {
     EMAIL,
